デモ画面のユースケースをカスタムフックに外だし
