+++ example-app/SantokuApp/src/navigation/DemoStackNav.tsx
@@ -17,7 +17,7 @@ import {
   MessageScreen,
   ButtonScreen,
   SnackbarScreen,
-  LocalAuthnScreen,
+  LocalAuthenticationScreen,
 } from 'screens';
 
 import {useCloseThisNavigatorButton} from './useCloseThisNavigatorButton';
@@ -50,7 +50,7 @@ export const Screen: React.FC = () => {
       <nav.Screen {...MessageScreen} />
       <nav.Screen {...ButtonScreen} />
       <nav.Screen {...SnackbarScreen} />
-      <nav.Screen {...LocalAuthnScreen} />
+      <nav.Screen {...LocalAuthenticationScreen} />
     </nav.Navigator>
   );
 };
