+++ example-app/SantokuApp/src/screens/demo/DemoScreen.tsx
@@ -10,7 +10,7 @@ import {ConfigScreen} from './config';
 import {ErrorCaseScreen} from './error';
 import {AppInfoScreen} from './info';
 import {InstructionsScreen} from './instructions';
-import {LocalAuthnScreen} from './local-authn';
+import {LocalAuthenticationScreen} from './local-authn';
 import {LogScreen} from './log';
 import {MessageScreen} from './message';
 import {SnackbarScreen} from './snackbar';
@@ -54,7 +54,7 @@ const demoScreenList = [
   },
   {
     title: 'Local Authentication',
-    to: LocalAuthnScreen.name,
+    to: LocalAuthenticationScreen.name,
   },
 ];
