+++ example-app/SantokuApp/src/screens/demo/local-authn/index.ts
@@ -1,1 +1,1 @@
-export * from './LocalAuthnScreen';
+export * from './LocalAuthenticationScreen';
