+++ example-app/SantokuApp/src/screens/demo/local-authn/useLocalAuthentication.tsx
@@ -1,11 +1,6 @@
-import React from 'react';
-import {StyleSheet, View} from 'react-native';
-import {Button} from 'react-native-elements';
+import {LocalAuthentication} from 'framework';
 
-import {LocalAuthentication} from '../../../framework';
-
-const ScreenName = 'Local Authentication';
-const Screen = () => {
+export const useLocalAuthentication = () => {
   const isEnrolled = async () => {
     const isEnrolled: boolean = await LocalAuthentication.isEnrolled();
     alert(isEnrolled);
@@ -44,28 +39,12 @@ const Screen = () => {
     });
     alert(JSON.stringify(result));
   };
-
-  return (
-    <View style={styles.container}>
-      <Button onPress={isEnrolled} title="デバイスの設定チェック" />
-      <Button onPress={isFingerPrintSupported} title="指紋認証のサポートのチェック" />
-      <Button onPress={isFacialSupported} title="顔認証のサポートのチェック" />
-      <Button onPress={isIrisSupported} title="虹彩認証のサポートのチェック" />
-      <Button onPress={localAuthentication1} title="認証の実行 Option指定3つ" />
-      <Button onPress={localAuthentication2} title="認証の実行 Option全指定" />
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'space-evenly',
-  },
-});
-
-export const LocalAuthnScreen = {
-  name: ScreenName,
-  component: Screen,
+  return {
+    isEnrolled,
+    isFingerPrintSupported,
+    isFacialSupported,
+    isIrisSupported,
+    localAuthentication1,
+    localAuthentication2,
+  };
 };
